feat: eliminates a controller that only had a static method
